Renamed the Processors namespace to be singular.
